requirements.txt
numpy==2.4.2
pillow==12.1.1
tifffile==2026.3.3
